pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["paridhi7 <paridhi.agarwal7@yahoo.in>"]
description = ""
name = "chess-bot-1"
version = "0.1.0"
[tool.poetry.dependencies]
cairosvg = "^2.4"
drawsvg = "^1.5"
ipython = "^7.14"
python = "^3.7"
python-chess = "^0.31.0"
tweepy = "^3.8"
[tool.poetry.dev-dependencies]
